03.20 구동 오류 해결

+++ sixbeam/src/main/java/com/erpproject/sixbeam/ac/service/SalesService.java
@@ -38,7 +38,7 @@ public class SalesService {
         return this.saleRepository.findBySaleBillingSt(false);
     }
 
-    public int getSaleList(String accountCd) {
+    public long getSaleList(String accountCd) {
         return estimateService.getAccountTotal(accountCd);
     }
